Analysis of wagyu source code by PVS-Studio
To demonstrate the abilities of PVS-Studio analyzer, we decided to find several bugs.
I would like to suggest a variant of the way to fix the error, detected with the help of V501 diagnostic: There are identical sub-expressions to the left and to the right of the '-' operator: pt2.x - pt2.x

PVS-Studio is a tool for bug detection in the source code of programs, written in C, C++ and C#. It works in Windows and Linux environment. https://www.viva64.com/en/pvs-studio/

We suggests having a look at the emails, sent from @pvs-studio.com.